mv: add io module and move swc things there
