+++ jaxley/__init__.py
@@ -12,3 +12,4 @@ from jaxley.modules import *
 from jaxley.optimize import ParamTransform
 from jaxley.stimulus import datapoint_to_step_currents, step_current
 from jaxley.utils.misc_utils import deprecated, deprecated_kwargs
+from jaxley.io.swc import read_swc
