+++ jaxley/modules/__init__.py
@@ -3,6 +3,6 @@
 
 from jaxley.modules.base import Module
 from jaxley.modules.branch import Branch
-from jaxley.modules.cell import Cell, read_swc
+from jaxley.modules.cell import Cell
 from jaxley.modules.compartment import Compartment
 from jaxley.modules.network import Network
